+++ test/add_column_test.rb
@@ -2,45 +2,17 @@ require_relative "test_helper"
 
 class AddColumnTest < Minitest::Test
   def test_default
-    if mysql? || mariadb?
-      assert_safe AddColumnDefault
-    else
-      with_target_version(10) do
-        assert_unsafe AddColumnDefault
-      end
-    end
-  end
-
-  def test_default_database_specific_versions
-    skip unless postgresql?
-    StrongMigrations.target_postgresql_version = "10"
-    assert_unsafe AddColumnDefault
-  ensure
-    StrongMigrations.target_postgresql_version = nil
+    assert_safe AddColumnDefault
   end
 
   def test_default_null
     skip unless postgresql?
-
-    with_target_version(10) do
-      assert_unsafe AddColumnDefaultNull, "Adding a column with a null default"
-    end
+    assert_safe AddColumnDefaultNull
   end
 
   def test_default_not_null
     skip unless postgresql?
-
-    with_target_version(10) do
-      assert_unsafe AddColumnDefaultNotNull, "Then add the NOT NULL constraint"
-    end
-  end
-
-  def test_default_safe_latest
-    skip unless postgresql? || mysql? || mariadb?
-
-    with_target_version(postgresql? ? 11 : (mysql? ? "8.0.12" : "10.3.2")) do
-      assert_safe AddColumnDefault
-    end
+    assert_unsafe AddColumnDefaultNotNull, "Then add the NOT NULL constraint"
   end
 
   def test_default_safe