+++ test/change_column_null_test.rb
@@ -12,33 +12,19 @@ class ChangeColumnNullTest < Minitest::Test
   def test_constraint
     skip unless postgresql?
 
-    with_target_version(12) do
-      assert_safe ChangeColumnNullConstraint
-    end
+    assert_safe ChangeColumnNullConstraint
   end
 
   def test_constraint_unvalidated
     skip unless postgresql?
 
-    with_target_version(12) do
-      assert_unsafe ChangeColumnNullConstraintUnvalidated
-    end
+    assert_unsafe ChangeColumnNullConstraintUnvalidated
   end
 
   def test_constraint_default
     skip unless postgresql?
 
-    with_target_version(12) do
-      assert_unsafe ChangeColumnNullConstraintDefault
-    end
-  end
-
-  def test_constraint_before_12
-    skip unless postgresql?
-
-    with_target_version(11) do
-      assert_unsafe ChangeColumnNullConstraint
-    end
+    assert_unsafe ChangeColumnNullConstraintDefault
   end
 
   def test_default
@@ -48,17 +34,13 @@ class ChangeColumnNullTest < Minitest::Test
   def test_constraint_methods
     skip unless postgresql?
 
-    with_target_version(12) do
-      assert_safe ChangeColumnNullConstraintMethods
-    end
+    assert_safe ChangeColumnNullConstraintMethods
   end
 
   def test_quoted
     skip unless postgresql?
 
-    with_target_version(12) do
-      assert_safe ChangeColumnNullQuoted
-    end
+    assert_safe ChangeColumnNullQuoted
   end
 
   def test_mysql_non_strict_mode