+++ test/multiple_databases_test.rb
@@ -4,9 +4,10 @@ class MultipleDatabasesTest < Minitest::Test
   def test_target_version
     skip unless postgresql?
 
-    with_target_version({primary: 10, animals: 11}) do
+    with_target_version({primary: 12, animals: 16}) do
       with_database(:primary) do
-        assert_unsafe AddColumnDefault
+        # TODO use new check
+        # assert_unsafe AddColumnDefault
       end
       with_database(:animals) do
         assert_safe AddColumnDefault
@@ -16,7 +17,7 @@ class MultipleDatabasesTest < Minitest::Test
 
   def test_target_version_unconfigured
     error = assert_raises(StrongMigrations::Error) do
-      with_target_version({primary: 10}) do
+      with_target_version({primary: 12}) do
         with_database(:animals) do
           assert_safe AddColumnDefault
         end