+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.9.0 (unreleased)
 
+- Improved install generator for Trilogy
 - Fixed charset check for MariaDB 11.4
 - Dropped support for Ruby < 3.1 and Active Record < 6.1